added basic user update tests

+++ database/migrations/2023_11_11_184139_change_user_column_name_in_unread_messages_table.php
@@ -2,6 +2,7 @@
 
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
@@ -12,8 +13,11 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('unread_messages', function (Blueprint $table) {
-            $table->dropForeign('unread_messages_user_id_foreign');
-            $table->dropColumn('user_id');
+            if (DB::getDriverName() !== 'sqlite') {
+                $table->dropForeign('unread_messages_user_id_foreign');
+                $table->dropColumn('user_id');
+            }
+
             $table->foreignId('unread_by_id')->constrained('users')->onDelete('cascade')->onUpdate('cascade');
         });
     }
@@ -24,8 +28,10 @@ return new class extends Migration
     public function down(): void
     {
         Schema::table('unread_messages', function (Blueprint $table) {
-            $table->dropForeign('unread_messages_unread_by_id_foreign');
-            $table->dropColumn('unread_by_id');
+            if (DB::getDriverName() !== 'sqlite') {
+                $table->dropForeign('unread_messages_unread_by_id_foreign');
+                $table->dropColumn('unread_by_id');
+            }
             $table->foreignId('user_id')->constrained('users')->onDelete('cascade')->onUpdate('cascade');
         });
     }
